# [REF] mrp: split productions