--- a/addons/mrp/tests/__init__.py
+++ b/addons/mrp/tests/__init__.py
@@ -14,3 +14,4 @@ from . import test_traceability
 from . import test_multicompany
 from . import test_backorder
 from . import test_smp
+from . import test_performance