--- a/addons/mrp/wizard/stock_assign_serial_numbers.py
+++ b/addons/mrp/wizard/stock_assign_serial_numbers.py
@@ -16,6 +16,7 @@ class StockAssignSerialNumbers(models.TransientModel):
     produced_qty = fields.Float('Produced Quantity', digits='Product Unit of Measure')
     show_apply = fields.Boolean(help="Technical field to show the Apply button")
     show_backorders = fields.Boolean(help="Technical field to show the Create Backorder and No Backorder buttons")
+    multiple_lot_components_names = fields.Text(help="Names of components with multiple lots, used to show warning")
 
     def generate_serial_numbers_production(self):
         if self.next_serial_number and self.next_serial_count: