--- a/addons/mrp/wizard/stock_assign_serial_numbers.xml
+++ b/addons/mrp/wizard/stock_assign_serial_numbers.xml
@@ -34,7 +34,7 @@
                     <group>
                         <field name="expected_qty" readonly="True"/>
                     </group>
-                    <p class="o_form_label oe_inline" attrs="{'invisible': [('show_backorders', '=', False)]}">
+                    <p col="1" class="o_form_label oe_inline" attrs="{'invisible': [('show_backorders', '=', False)]}">
                         You have entered less serial numbers than the quantity to produce.<br/>
                         Create a backorder if you expect to process the remaining quantities later.<br/>
                         Do not create a backorder if you will not process the remaining products.