--- a/addons/mrp/wizard/stock_assign_serial_numbers.xml
+++ b/addons/mrp/wizard/stock_assign_serial_numbers.xml
@@ -25,6 +25,13 @@
                 <group>
                     <field name="serial_numbers" placeholder="copy paste a list and/or use Generate"/>
                 </group>
+                <field name="multiple_lot_components_names" invisible="1"/>
+                <group col="1">
+                    <p class="o_form_label oe_inline" attrs="{'invisible': [('multiple_lot_components_names', '=', False)]}">
+                        Note that components <field name="multiple_lot_components_names" readonly="True"/> have multiple lot reservations.<br/>
+                        Do you want to confirm anyway ?
+                    </p>
+                </group>
                 <field name="show_apply" invisible="1" />
                 <field name="show_backorders" invisible="1" />
                 <group>